+++ crates/icydb-schema/src/node/index.rs
@@ -10,7 +10,6 @@ use std::{
 
 #[derive(Clone, Debug, Serialize)]
 pub struct Index {
-    pub store: &'static str,
     pub fields: &'static [&'static str],
 
     #[serde(default, skip_serializing_if = "Not::not")]
@@ -29,9 +28,9 @@ impl Display for Index {
         let fields = self.fields.join(", ");
 
         if self.unique {
-            write!(f, "UNIQUE {}({})", self.store, fields)
+            write!(f, "UNIQUE ({fields})")
         } else {
-            write!(f, "{}({})", self.store, fields)
+            write!(f, "({fields})")
         }
     }
 }
@@ -42,20 +41,7 @@ impl MacroNode for Index {
     }
 }
 
-impl ValidateNode for Index {
-    fn validate(&self) -> Result<(), ErrorTree> {
-        let mut errs = ErrorTree::new();
-        let schema = schema_read();
-
-        // store
-        match schema.cast_node::<IndexStore>(self.store) {
-            Ok(_) => {}
-            Err(e) => errs.add(e),
-        }
-
-        errs.result()
-    }
-}
+impl ValidateNode for Index {}
 
 impl VisitableNode for Index {
     fn route_key(&self) -> String {